+++ test/test_mongo_controller_audit_versioned.py
@@ -6,7 +6,7 @@ import pytest
 from layaberr import ModelCouldNotBeFound
 
 import layabase
-import layabase._database_mongo
+import layabase.mongo
 from layabase.testing import mock_mongo_audit_datetime
 
 
@@ -20,9 +20,9 @@ def controller():
     class TestCollection:
         __collection_name__ = "test"
 
-        key = layabase._database_mongo.Column(str, is_primary_key=True)
-        mandatory = layabase._database_mongo.Column(int, is_nullable=False)
-        optional = layabase._database_mongo.Column(str)
+        key = layabase.mongo.Column(str, is_primary_key=True)
+        mandatory = layabase.mongo.Column(int, is_nullable=False)
+        optional = layabase.mongo.Column(str)
 
     return layabase.CRUDController(TestCollection, audit=True)
 
@@ -32,8 +32,8 @@ def controller_versioned():
     class TestCollectionVersioned:
         __collection_name__ = "test_versioned"
 
-        key = layabase._database_mongo.Column(str, is_primary_key=True)
-        enum_fld = layabase._database_mongo.Column(EnumTest)
+        key = layabase.mongo.Column(str, is_primary_key=True)
+        enum_fld = layabase.mongo.Column(EnumTest)
 
     return layabase.CRUDController(TestCollectionVersioned, audit=True, history=True)