+++ test/test_mongo_controller_dict.py
@@ -6,7 +6,7 @@ import pytest
 from layaberr import ValidationFailed
 
 import layabase
-import layabase._database_mongo
+import layabase.mongo
 
 
 class EnumTest(enum.Enum):
@@ -19,13 +19,11 @@ def controller():
     class TestCollection:
         __collection_name__ = "test"
 
-        key = layabase._database_mongo.Column(str, is_primary_key=True)
-        dict_col = layabase._database_mongo.DictColumn(
+        key = layabase.mongo.Column(str, is_primary_key=True)
+        dict_col = layabase.mongo.DictColumn(
             fields={
-                "first_key": layabase._database_mongo.Column(
-                    EnumTest, is_nullable=False
-                ),
-                "second_key": layabase._database_mongo.Column(int, is_nullable=False),
+                "first_key": layabase.mongo.Column(EnumTest, is_nullable=False),
+                "second_key": layabase.mongo.Column(int, is_nullable=False),
             },
             is_nullable=False,
         )