+++ test/test_mongo_controller_failures.py
@@ -1,7 +1,7 @@
 import pytest
 
 import layabase
-import layabase._database_mongo
+import layabase.mongo
 
 
 def test_namespace_method_without_connecting_to_database():
@@ -11,7 +11,7 @@ def test_namespace_method_without_connecting_to_database():
     class TestCollection:
         __collection_name__ = "test"
 
-        id = layabase._database_mongo.Column()
+        id = layabase.mongo.Column()
 
     with pytest.raises(layabase.ControllerModelNotSet) as exception_info:
         layabase.CRUDController(TestCollection).namespace(TestNamespace)
@@ -25,7 +25,7 @@ def test_counters_table_name_is_forbidden():
     class TestCollection:
         __collection_name__ = "counters"
 
-        key = layabase._database_mongo.Column(str)
+        key = layabase.mongo.Column(str)
 
     with pytest.raises(Exception) as exception_info:
         layabase.load("mongomock", [layabase.CRUDController(TestCollection)])
@@ -37,7 +37,7 @@ def test_audit_table_name_is_forbidden():
     class TestCollection:
         __collection_name__ = "audit"
 
-        key = layabase._database_mongo.Column(str)
+        key = layabase.mongo.Column(str)
 
     with pytest.raises(Exception) as exception_info:
         layabase.load("mongomock", [layabase.CRUDController(TestCollection)])
@@ -49,7 +49,7 @@ def test_audit_prefixed_table_name_is_forbidden():
     class TestCollection:
         __collection_name__ = "audit_toto"
 
-        key = layabase._database_mongo.Column(str)
+        key = layabase.mongo.Column(str)
 
     with pytest.raises(Exception) as exception_info:
         layabase.load("mongomock", [layabase.CRUDController(TestCollection)])
@@ -61,7 +61,7 @@ def test_get_method_without_connecting_to_database():
     class TestCollection:
         __collection_name__ = "test"
 
-        id = layabase._database_mongo.Column()
+        id = layabase.mongo.Column()
 
     with pytest.raises(layabase.ControllerModelNotSet) as exception_info:
         layabase.CRUDController(TestCollection).get({})
@@ -75,7 +75,7 @@ def test_post_method_without_connecting_to_database():
     class TestCollection:
         __collection_name__ = "test"
 
-        id = layabase._database_mongo.Column()
+        id = layabase.mongo.Column()
 
     with pytest.raises(layabase.ControllerModelNotSet) as exception_info:
         layabase.CRUDController(TestCollection).post({})
@@ -89,7 +89,7 @@ def test_post_many_method_without_connecting_to_database():
     class TestCollection:
         __collection_name__ = "test"
 
-        id = layabase._database_mongo.Column()
+        id = layabase.mongo.Column()
 
     with pytest.raises(layabase.ControllerModelNotSet) as exception_info:
         layabase.CRUDController(TestCollection).post_many([])
@@ -103,7 +103,7 @@ def test_put_method_without_connecting_to_database():
     class TestCollection:
         __collection_name__ = "test"
 
-        id = layabase._database_mongo.Column()
+        id = layabase.mongo.Column()
 
     with pytest.raises(layabase.ControllerModelNotSet) as exception_info:
         layabase.CRUDController(TestCollection).put({})
@@ -117,7 +117,7 @@ def test_delete_method_without_connecting_to_database():
     class TestCollection:
         __collection_name__ = "test"
 
-        id = layabase._database_mongo.Column()
+        id = layabase.mongo.Column()
 
     with pytest.raises(layabase.ControllerModelNotSet) as exception_info:
         layabase.CRUDController(TestCollection).delete({})
@@ -131,7 +131,7 @@ def test_audit_method_without_connecting_to_database():
     class TestCollection:
         __collection_name__ = "test"
 
-        id = layabase._database_mongo.Column()
+        id = layabase.mongo.Column()
 
     with pytest.raises(layabase.ControllerModelNotSet) as exception_info:
         layabase.CRUDController(TestCollection).get_audit({})
@@ -145,7 +145,7 @@ def test_model_description_method_without_connecting_to_database():
     class TestCollection:
         __collection_name__ = "test"
 
-        id = layabase._database_mongo.Column()
+        id = layabase.mongo.Column()
 
     with pytest.raises(layabase.ControllerModelNotSet) as exception_info:
         layabase.CRUDController(TestCollection).get_model_description()