+++ test/test_mongo_controller_list.py
@@ -5,7 +5,7 @@ import flask_restplus
 import pytest
 
 import layabase
-import layabase._database_mongo
+import layabase.mongo
 
 
 class EnumTest(enum.Enum):
@@ -18,20 +18,16 @@ def controller():
     class TestCollection:
         __collection_name__ = "test"
 
-        key = layabase._database_mongo.Column(is_primary_key=True)
-        list_field = layabase._database_mongo.ListColumn(
-            layabase._database_mongo.DictColumn(
+        key = layabase.mongo.Column(is_primary_key=True)
+        list_field = layabase.mongo.ListColumn(
+            layabase.mongo.DictColumn(
                 fields={
-                    "first_key": layabase._database_mongo.Column(
-                        EnumTest, is_nullable=False
-                    ),
-                    "second_key": layabase._database_mongo.Column(
-                        int, is_nullable=False
-                    ),
+                    "first_key": layabase.mongo.Column(EnumTest, is_nullable=False),
+                    "second_key": layabase.mongo.Column(int, is_nullable=False),
                 }
             )
         )
-        bool_field = layabase._database_mongo.Column(bool)
+        bool_field = layabase.mongo.Column(bool)
 
     controller = layabase.CRUDController(TestCollection)
     layabase.load("mongomock", [controller])